+++ src/main/kotlin/spock/adb/actions/RestartAppAction.kt
@@ -1,10 +1,11 @@
 package spock.adb.actions
 
-import com.android.ddmlib.IDevice
-import spock.adb.AdbController
+import com.intellij.openapi.project.Project
+import spock.adb.SpockAdbService
+import spock.adb.device.ConnectedDevice
 
-class RestartAppAction : BaseAction() {
-    override fun performAction(controller: AdbController, device: IDevice) {
-        controller.restartApp(device)
-    }
+class RestartAppAction : DeviceAwareAction() {
+    override val baseDescription = "Force-stop the app and launch it again"
+    override fun perform(project: Project, device: ConnectedDevice) =
+        SpockAdbService.getInstance(project).controller.restartApp(device.device)
 }